+++ CHANGELOG.md
@@ -7,8 +7,12 @@
 * Easter egg quest  #87
 * Sorcerer enemy class  #88
 
+### Changed
+* Remember unlocked quests and todo list order #89
+
 ### Fixed
 * Reach level 50 and 100 unlock and reward 4128f75
+* Properly report raise class levels quest progress e7d73f9
 
 ## [0.6.0](https://github.com/facundoolano/rpg-cli/releases/tag/0.6.0) - 2021-08-04
 ### Added